Added gcc printf format checking via __attribute__ (if using gcc). This spotted a myriad of minor bugs (mostly in error handling code), even including some 10+ years old. Now fixed.
git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2359 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1

+++ src/gap4/tkEditor.c
@@ -1829,7 +1829,7 @@ static int EditorWidgetCmd(ClientData clientData, Tcl_Interp *interp,
                 redisplaySequences(ed->xx, 1);
             }
             vTcl_SetResult(interp, "%d",
-                           DBI_flags(ed->xx) & DB_ACCESS_UPDATE);
+                           (int)(DBI_flags(ed->xx) & DB_ACCESS_UPDATE));
         }
     } else if ('l' == *argv[1] && strcmp(argv[1], "list_confidence") == 0) {
 	int left, right, info_only;
